+++ lexicmap/cmd/search.go
@@ -595,7 +595,7 @@ func init() {
 		formatFlagUsage(`Maximum evalue of a HSP segment.`))
 
 	mapCmd.Flags().BoolP("debug", "", false,
-		formatFlagUsage(`Print debug information.`))
+		formatFlagUsage(`Print debug information, including a progress bar. (recommended when searching with one query).`))
 
 	mapCmd.SetUsageTemplate(usageTemplate("-d <index path> [query.fasta.gz ...] [-o query.tsv.gz]"))